Move `Quiz` to it's own file in production.
